// Cognexia-Frontend-CRM_Frontend/backend/modules/05-inventory/src/entities/StockTransaction.d.ts
import { TransactionType, TransactionReason } from '../enums';
import { InventoryItem } from './InventoryItem';
export declare class StockTransaction {
    id: string;
    itemId: string;
    type: TransactionType;
    quantity: number;
    balanceAfter: number;
    reason: TransactionReason;
    locationCode?: string;
    batchNumber?: string;
    serialNumber?: string;
    expiryDate?: Date;
    notes?: string;
    referenceId?: string;
    performedBy: string;
    unitCost?: number;
    totalCost?: number;
    documentNumber?: string;
    supplierReference?: string;
    customerReference?: string;
    createdAt: Date;
    item: InventoryItem;
    calculateTotalCost(): void;
    isInbound(): boolean;
    isOutbound(): boolean;
    getTransactionValue(): number;
    isAdjustment(): boolean;
    isCycleCount(): boolean;
    isPurchaseRelated(): boolean;
    isSaleRelated(): boolean;
    isProductionRelated(): boolean;
    getImpactDescription(): string;
    toJSON(): {
        id: string;
        itemId: string;
        type: TransactionType;
        quantity: number;
        balanceAfter: number;
        reason: TransactionReason;
        locationCode: string | undefined;
        batchNumber: string | undefined;
        serialNumber: string | undefined;
        expiryDate: Date | undefined;
        notes: string | undefined;
        referenceId: string | undefined;
        performedBy: string;
        unitCost: number | undefined;
        totalCost: number | undefined;
        documentNumber: string | undefined;
        supplierReference: string | undefined;
        customerReference: string | undefined;
        createdAt: Date;
        isInbound: boolean;
        isOutbound: boolean;
        transactionValue: number;
        impactDescription: string;
    };
}
//# sourceMappingURL=StockTransaction.d.ts.map